feat: Sanitize package name arguments
My initial issue was that I wnated to be able to write `npm-compare */` which in bash would pass the list of all subdirectories to the tool, but all end with a `/`.
By doing that the toll still made all those API requests to npms.io, which I think we can avoid.
So with this change npm-compare tries to
- considers all white space in arguments as separators (multiple as a single one) between package names
- removes some invalid chars from beginning and end of arguments (including `/` and `\` from directory list)
- filters the result using `validate-npm-package-name` and reports the skipped ones to stderr (incl the reason that lib provides)

I ran both `yarn add validate-npm-package-name` and `npm i validate-npm-package-name` to update both lockfiles.

https://github.com/npm/validate-npm-package-name
https://github.com/sanchitnevgi/npm-compare/issues/14